+++ pkgs/unified_analytics/lib/src/initializer.dart
@@ -17,7 +17,6 @@ class Initializer {
   final Directory homeDirectory;
   final int toolsMessageVersion;
   bool firstRun = false;
-  final bool pddFlag;
 
   /// Responsibe for the initialization of the files
   /// necessary for analytics reporting
@@ -33,7 +32,6 @@ class Initializer {
     required this.tool,
     required this.homeDirectory,
     required this.toolsMessageVersion,
-    required this.pddFlag,
   });
 
   /// Creates the text file that will contain the client ID
@@ -118,14 +116,14 @@ class Initializer {
     // Begin initialization checks for the session file
     final File sessionFile = fs.file(
         p.join(homeDirectory.path, kDartToolDirectoryName, kSessionFileName));
-    if (!sessionFile.existsSync() && !pddFlag) {
+    if (!sessionFile.existsSync()) {
       createSessionFile(sessionFile: sessionFile);
     }
 
     // Begin initialization checks for the log file to persist events locally
     final File logFile = fs
         .file(p.join(homeDirectory.path, kDartToolDirectoryName, kLogFileName));
-    if (!logFile.existsSync() && !pddFlag) {
+    if (!logFile.existsSync()) {
       createLogFile(logFile: logFile);
     }
   }